feat! Context-awareness
Update client implementation to accept Contexts as a parameter. The
context is then passed to client-go directly.

The current implementation creates a separate Context for each api
call, which erases any context information that can be passed from the
calling function or goroutine. This prevents manifestival from
supporting timeouts, cancellation, deadlines, or graceful termination.
Wiring a Context object through to client-go, per the golang convention
of passing it as the first argument, fixes this issue.

BREAKING CHANGE: This update alters the Client implementation to align
with its parent interface breaking change in manifestival/manifestival.

Signed-off-by: Adam Kaplan <adam@adambkaplan.com>